fix: examples
