+++ examples/4_rag/README.md
@@ -1,3 +1,3 @@
 # RAG
 
-For basic rag application, refer to [rag_langchain_demo.py](../rag_langchain_demo.py)
+For basic rag application, refer to [rag_langchain_demo.py](rag_langchain_demo.py)
