+++ examples/README_zh_CN.md
@@ -1,0 +1,13 @@
+# Open-Retrievals 示例
+
+
+## 向量模型
+
+
+## 检索
+
+
+## 重排
+
+
+## 问答
